+++ noodles-cram/src/data_container/compression_header/preservation_map/substitution_matrix/builder.rs
@@ -1,3 +1,5 @@
+use noodles_sam::RecordExt;
+
 use crate::{record::Feature, Record};
 
 use super::{Base, Histogram, SubstitutionMatrix};