+++ noodles-cram/src/data_container/slice.rs
@@ -6,7 +6,7 @@ pub use self::{builder::Builder, header::Header};
 use std::io::{self, Cursor};
 
 use noodles_fasta as fasta;
-use noodles_sam as sam;
+use noodles_sam::{self as sam, RecordExt};
 
 use super::CompressionHeader;
 use crate::{