+++ noodles-cram/src/indexer.rs
@@ -1,6 +1,7 @@
 use std::{cmp, collections::HashMap, fs::File, io, path::Path};
 
 use noodles_bam as bam;
+use noodles_sam::RecordExt;
 
 use super::{crai, data_container, Reader};