+++ noodles-cram/src/reader/record.rs
@@ -9,7 +9,7 @@ use std::{
 
 use byteorder::ReadBytesExt;
 use noodles_bam as bam;
-use noodles_sam as sam;
+use noodles_sam::{self as sam, RecordExt};
 
 use super::num::read_itf8;
 use crate::{