+++ noodles-cram/src/writer/record.rs
@@ -9,7 +9,7 @@ use std::{
 use byteorder::WriteBytesExt;
 
 use noodles_bam as bam;
-use noodles_sam as sam;
+use noodles_sam::{self as sam, RecordExt};
 
 use super::num::write_itf8;
 use crate::{